requirements.txt
aiogram==3.19.0
python-dotenv==1.1.0
SQLAlchemy==2.0.35
aiohttp==3.10.5
aiosqlite==0.20.0
asyncpg==0.29.0